move the useFormatDistance to a parser file
