+++ src/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from './useFormatDates'
-export * from './useFormatDistance'
