+++ src/libs/parsers/index.ts
@@ -1,2 +1,3 @@
 export * from './category'
 export * from './getDisplayPrice'
+export * from './formatDistance'
